+++ entry.c
@@ -92,7 +92,7 @@ static int create_file(const char *path, unsigned int mode)
 void *read_blob_entry(const struct cache_entry *ce, size_t *size)
 {
 	enum object_type type;
-	unsigned long ul;
+	size_t ul;
 	void *blob_data = odb_read_object(the_repository->objects, &ce->oid,
 					  &type, &ul);
